# Backend/Verification/baseline-openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "BTech",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/sections/{sectionId}/students/{studentId}": {
      "delete": {
        "tags": [
          "SectionAssignments"
        ],
        "parameters": [
          {
            "name": "sectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/student-parents/{studentId}": {
      "get": {
        "tags": [
          "StudentParent"
        ],
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "StudentParent"
        ],
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentParent"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentParent"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StudentParent"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/students/{studentId}": {
      "get": {
        "tags": [
          "Students"
        ],
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Students"
        ],
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStudentDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStudentDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStudentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "StudentParent": {
        "type": "object",
        "properties": {
          "parentId": {
            "type": "integer",
            "format": "int64"
          },
          "studentId": {
            "type": "integer",
            "format": "int64"
          },
          "fatherName": {
            "type": "string",
            "nullable": true
          },
          "fatherMobile": {
            "type": "string",
            "nullable": true
          },
          "fatherEmail": {
            "type": "string",
            "nullable": true
          },
          "fatherOccupation": {
            "type": "string",
            "nullable": true
          },
          "motherName": {
            "type": "string",
            "nullable": true
          },
          "motherMobile": {
            "type": "string",
            "nullable": true
          },
          "motherEmail": {
            "type": "string",
            "nullable": true
          },
          "motherOccupation": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateStudentDto": {
        "required": [
          "fullName",
          "studentCode"
        ],
        "type": "object",
        "properties": {
          "collegeId": {
            "minimum": 1,
            "type": "integer",
            "format": "int64"
          },
          "studentCode": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "fullName": {
            "maxLength": 150,
            "minLength": 0,
            "type": "string"
          },
          "gender": {
            "maxLength": 20,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "email": {
            "maxLength": 150,
            "minLength": 0,
            "type": "string",
            "format": "email",
            "nullable": true
          },
          "mobile": {
            "maxLength": 20,
            "minLength": 0,
            "pattern": "^[0-9+() -]*$",
            "type": "string",
            "nullable": true
          },
          "bloodGroup": {
            "maxLength": 10,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "address": {
            "maxLength": 500,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "courseId": {
            "minimum": 1,
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "branchId": {
            "minimum": 1,
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "academicYearId": {
            "minimum": 1,
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "maximum": 1,
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Enter JWT token as: Bearer {token}",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}